docker/spark-driver/drf_app/requirements.txt
Django
djangorestframework
pyspark
elasticsearch

